implement /cats route.
